requirements.txt
ipykernel==6.29.0
lime==0.2.0.1
matplotlib==3.7.1
numpy==1.26.4
pandas==2.0.1
pathlib==1.0.1
pillow==10.2.0
rembg==2.0.54
scikit-learn==1.2.2
scikit-image==0.22.0
seaborn==0.11.2
statsmodels==0.14.0
streamlit==1.31.1
tensorflow==2.13.1
tensorflow-metal==1.1.0